+++ src/app.module.ts
@@ -1,8 +1,7 @@
-import { MiddlewareConsumer, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { MailerModule } from '@nestjs-modules/mailer';
 import { EjsAdapter } from '@nestjs-modules/mailer/dist/adapters/ejs.adapter';
-import { CheckAuthenticatedMiddleware } from './middlewares/check-authenticated/check-authenticated.middleware';
 import { configuration } from './configs/configuration';
 import { CatsModule } from './routes/cats/cats.module';
 
@@ -36,9 +35,4 @@ import { CatsModule } from './routes/cats/cats.module';
   controllers: [],
   providers: [],
 })
-export class AppModule {
-  configure(consumer: MiddlewareConsumer) {
-    // Uncomment to enable middleware
-    // consumer.apply(CheckAuthenticatedMiddleware).forRoutes('*');
-  }
-}
+export class AppModule {}