+++ src/decorators/authorize.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+import { Roles } from 'src/guards/is-authorized/roles';
+
+export const Authorize = (...authorize: Roles[]) => SetMetadata('authorize', authorize);