+++ src/guards/is-authorized/is-authorized.guard.spec.ts
@@ -1,0 +1,7 @@
+import { IsAuthorizedGuard } from './is-authorized.guard';
+
+describe('IsAuthorizedGuard', () => {
+  it('should be defined', () => {
+    expect(new IsAuthorizedGuard()).toBeDefined();
+  });
+});