+++ src/middlewares/check-authenticated/check-authenticated.middleware.spec.ts
@@ -1,7 +1,0 @@
-import { CheckAuthenticatedMiddleware } from './check-authenticated.middleware';
-
-describe('CheckAuthenticatedMiddleware', () => {
-  it('should be defined', () => {
-    expect(new CheckAuthenticatedMiddleware()).toBeDefined();
-  });
-});