+++ src/types/session.d.ts
@@ -1,0 +1,12 @@
+import 'express-session';
+
+declare module 'express-session' {
+  interface Session {
+    user?: {
+      isLogged: boolean;
+      isAdmin?: boolean;
+
+      // Declare here your session data
+    };
+  }
+}